权限控制完成

+++ src/main/java/com/honji/order/controller/SalesPlanDetailsController.java
@@ -1,12 +1,15 @@
 package com.honji.order.controller;
 
 
-import com.honji.order.entity.CashBalance;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.honji.order.entity.Area;
+import com.honji.order.entity.DetailsProposal;
 import com.honji.order.entity.SalesPlan;
 import com.honji.order.entity.SalesPlanDetails;
 import com.honji.order.mapper.SalesPlanMapper;
 import com.honji.order.model.DataGridResult;
-import com.honji.order.service.ICashBalanceService;
+import com.honji.order.service.IAreaService;
+import com.honji.order.service.IDetailsProposalService;
 import com.honji.order.service.ISalesPlanDetailsService;
 import com.honji.order.service.ISalesPlanService;
 import lombok.extern.slf4j.Slf4j;
@@ -40,7 +43,10 @@ public class SalesPlanDetailsController {
     @Autowired
     private ISalesPlanService salesPlanService;
     @Autowired
-    private ICashBalanceService cashBalanceService;
+    private IAreaService areaService;
+
+    @Autowired
+    private IDetailsProposalService detailsProposalService;
     @Autowired
     private SalesPlanMapper salesPlanMapper;
 
@@ -57,8 +63,8 @@ public class SalesPlanDetailsController {
             model.addAttribute("name", map.get("name"));
             model.addAttribute("createDate", LocalDate.now());
         }
-        List<CashBalance> shops = cashBalanceService.list();
-        model.addAttribute("shops", shops);
+        List<Area> areas = areaService.list();
+        model.addAttribute("areas", areas);
         model.addAttribute("jobNum", jobNum);
         return "sales-plan-details";
     }
@@ -78,15 +84,35 @@ public class SalesPlanDetailsController {
 
     @PostMapping("/save")
     @ResponseBody
-    public String save(@ModelAttribute SalesPlanDetails salesPlanDetails) {
+    public String save(@ModelAttribute SalesPlanDetails salesPlanDetails, @RequestParam boolean wantToNotify) {
+        if (wantToNotify) {
+            SalesPlan salesPlan = salesPlanMapper.selectById(salesPlanDetails.getPlanId());
+            String jobNum = salesPlanMapper.selectManagerByShop(salesPlan.getShopCode());
+            salesPlanMapper.notify(jobNum, "您有业绩下降原因上报，请查看！");
+        }
         salesPlanDetailsService.saveOrUpdate(salesPlanDetails);
         return salesPlanDetails.getId();
+    }
+
+    @PostMapping("/saveFeedback")
+    @ResponseBody
+    public void saveFeedback(@RequestParam String id, @RequestParam String feedback) {
+
+        salesPlanDetailsService.saveFeedback(id, feedback);
     }
 
 
     @PostMapping("/remove")
     @ResponseBody
     public boolean remove(@RequestParam String id) {
+        SalesPlanDetails salesPlanDetails =  salesPlanDetailsService.getById(id);
+        QueryWrapper<DetailsProposal> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("details_id", id);
+        List<DetailsProposal> proposals = detailsProposalService.list(queryWrapper);
+        //已经有方案或者有反馈不允许删除
+        if (proposals.size() > 0 || salesPlanDetails.getFeedback().length() > 0) {
+            return false;
+        }
         return salesPlanDetailsService.removeById(id);
     }
